requirements.txt
langchain
openai
dotenv
phoenix-tracing
rouge-score
pytest